Move base classes to own module
